+++ lib/pack_game/levels_hud/controls/widgets/word_composition_row.dart
@@ -10,6 +10,8 @@ import 'package:wbw_core/wbw_core.dart';
 import 'package:wbw_design_core/wbw_design_core.dart';
 import 'package:word_by_word_game/generated/l10n.dart';
 import 'package:word_by_word_game/pack_core/global_states/global_states.dart';
+import 'package:word_by_word_game/pack_game/levels_hud/controls/widgets/send_action_button.dart';
+import 'package:word_by_word_game/pack_game/levels_hud/controls/widgets/warning_notification.dart';
 import 'package:word_by_word_game/pack_game/mechanics/mechanics.dart';
 
 part 'word_composition_row_state.dart';
@@ -39,7 +41,6 @@ class WordCompositionRow extends HookWidget {
           if (levelState is! LiveLevelBlocState) return const SizedBox();
 
           return Column(
-            mainAxisSize: MainAxisSize.min,
             children: [
               LastWordText(latestWord: levelState.latestWord),
               Builder(
@@ -90,6 +91,19 @@ class WordCompositionRow extends HookWidget {
                     ],
                   );
                 },
+              ),
+              uiTheme.verticalBoxes.medium,
+              Expanded(
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    const Flexible(child: WarningNotification()),
+                    Padding(
+                      padding: EdgeInsets.only(left: spacing.extraLarge),
+                      child: const SendWordActionButton(),
+                    ),
+                  ],
+                ),
               ),
             ],
           );