+++ lib/pack_game/levels_hud/controls/widgets/word_composition_row_state.dart
@@ -56,6 +56,7 @@ class _WordCompositionState extends LifeState {
         rightPart: rightPartController.text,
       ),
     );
+
     _wordUpdatesController.add(newWord);
   }